+++ examples/applb_gwlb_fgtasg/outputs.tf
@@ -29,3 +29,19 @@ output "standard_lb_backend_pool_id" {
   description = "The backend ID of the Standard Load Balancer that servers connected with."
   value       = { for k, v in module.standard_load_balancer : k => v.backend_pool_id }
 }
+
+output "gwlb" {
+  value = module.gwlb
+}
+
+output "gwlb_extract" {
+  value = {
+    for k, v in var.fortigate_scaleset : k => {
+      for interface in v.network_interfaces : interface.name => {
+        gwlb_frontend_ip_configuration_id = try(interface.gwlb_key, null) != null ? module.gwlb[interface.gwlb_key].frontend_ip_config_id : ""
+        gwlb_backend_pool_id              = try(interface.gwlb_key, null) != null ? module.gwlb[interface.gwlb_key].backend_pool_ids[interface.gwlb_backend_key] : ""
+        gwlb_private_ip_address           = try(interface.gwlb_key, null) != null ? module.gwlb[interface.gwlb_key].private_ip_address : ""
+      }
+    }
+  }
+}